requirements.txt
# K-Scan Security Audit System Dependencies

# Core OpenAI Agents SDK
openai-agents>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client and Web Scraping
aiohttp>=3.9.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Validation and Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlmodel>=0.0.14
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2

# Supabase for content storage
supabase>=2.1.0
PyJWT>=2.8.0

# Environment and Configuration
python-dotenv>=1.0.0

# Security and Cryptography
cryptography>=41.0.0
requests>=2.31.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Optional: Browser automation (if not using Browserless.io)
# playwright>=1.40.0
# selenium>=4.15.0 

fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
sqlmodel
psycopg2-binary
alembic
sqlalchemy
asyncpg
celery
redis
flower
gevent 